+++ package.json
@@ -20,18 +20,13 @@
   "homepage": "https://gitlab.com/inezabonte/todo-frontend#readme",
   "dependencies": {
     "@ant-design/icons": "^4.5.0",
-    "@material-ui/core": "^4.11.3",
-    "@material-ui/icons": "^4.11.2",
-    "@material-ui/lab": "^4.0.0-alpha.57",
     "antd": "^4.12.3",
-    "formik": "^2.2.6",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
     "react-redux": "^7.2.2",
     "react-router-dom": "^5.2.0",
     "redux": "^4.0.5",
-    "redux-thunk": "^2.3.0",
-    "yup": "^0.32.8"
+    "redux-thunk": "^2.3.0"
   },
   "devDependencies": {
     "@babel/core": "^7.12.10",
