--- src/App.css
@@ -1,1 +1,0 @@
-@import '~antd/dist/antd.css';
