+++ src/components/App/style.css
@@ -1,0 +1,3 @@
+.todo-container{
+    margin-top: 50px;
+}
