+++ src/index.css
@@ -1,0 +1,5 @@
+@import '~antd/dist/antd.css';
+
+body{
+    background-color: rgb(245, 242, 242);
+}
